Move axes setup to coord classes
